// src/services/api/gen/types.gen.ts
export type ApprovalStatusFe1Enum = 'approved' | 'rejected' | 'pending' | 'need_information' | 'kyc_pending';
// ...
export type BusinessList = {
    readonly id: string;
    readonly name: string;
    readonly description: string | null;
    readonly logo: string | null;
    readonly contact_no: string;
    readonly tagline: string | null;
    readonly opening_time: string;
    readonly closing_time: string;
    readonly website: string | null;
    approval_status: ApprovalStatusFe1Enum;
    readonly remark: string;
    establishment: string;
};